MAINT: switch number of decays to long

+++ workspace/Gray/include/Gray/GammaRayTrace.h
@@ -12,13 +12,13 @@ class SourceList;
 
 class GammaRayTrace {
 public:
-    static int TraceSources(SourceList & sources,
-                            IntersectKdTree & tree,
-                            int num_decays,
-                            std::vector<Interaction> & interactions,
-                            size_t soft_max_interactions,
-                            GammaMaterial const * const default_material,
-                            bool log_nuclear_decays);
+    static long TraceSources(SourceList & sources,
+                             IntersectKdTree & tree,
+                             long num_decays,
+                             std::vector<Interaction> & interactions,
+                             size_t soft_max_interactions,
+                             GammaMaterial const * const default_material,
+                             bool log_nuclear_decays);
 private:
     static void TracePhoton(Photon &photon,
                             std::vector<Interaction> & interactions,
