+++ workspace/Gray/src/GammaRayTrace.cpp
@@ -109,15 +109,15 @@ void GammaRayTrace::TracePhoton(
     return;
 }
 
-int GammaRayTrace::TraceSources(SourceList & sources,
-                                IntersectKdTree & tree,
-                                int num_decays,
-                                std::vector<Interaction> & interactions,
-                                size_t soft_max_interactions,
-                                GammaMaterial const * const default_material,
-                                bool log_nuclear_decays)
+long GammaRayTrace::TraceSources(SourceList & sources,
+                                 IntersectKdTree & tree,
+                                 long num_decays,
+                                 std::vector<Interaction> & interactions,
+                                 size_t soft_max_interactions,
+                                 GammaMaterial const * const default_material,
+                                 bool log_nuclear_decays)
 {
-    for (int i = 0; i < num_decays; i++) {
+    for (long i = 0; i < num_decays; i++) {
         Source * source = sources.Decay();
         Isotope * isotope = source->GetIsotope();
         if (isotope == NULL) {
